fix: let lambda continue when sqs-messages blob is not found (#1306)

+++ src/RoadRegistry.BackOffice.Handlers.Sqs.Lambda/BlobMessageHandler.cs
@@ -26,16 +26,23 @@ public abstract class BlobMessageHandler : IMessageHandler
             var sw = Stopwatch.StartNew();
             messageMetadata.Logger?.LogInformation($"Downloading blob '{blobName}'");
             var blobMessageData = await _blobClient.GetBlobMessageAsync(blobName, cancellationToken);
-            messageMetadata.Logger?.LogInformation($"Downloaded blob '{blobName}'");
-            messageMetadata.Logger?.LogInformation($"TIMETRACKING blobmessagehandler: loading blobmessage took {sw.Elapsed}");
+            if (blobMessageData is not null)
+            {
+                messageMetadata.Logger?.LogInformation($"Downloaded blob '{blobName}'");
+                messageMetadata.Logger?.LogInformation($"TIMETRACKING blobmessagehandler: loading blobmessage took {sw.Elapsed}");
 
-            await HandleMessageDataAsync(blobMessageData, messageMetadata, cancellationToken);
+                await HandleMessageDataAsync(blobMessageData, messageMetadata, cancellationToken);
 
-            sw.Restart();
-            messageMetadata.Logger?.LogInformation($"Deleting blob '{blobName}'");
-            await _blobClient.DeleteBlobAsync(blobName, cancellationToken);
-            messageMetadata.Logger?.LogInformation($"TIMETRACKING blobmessagehandler: deleting blobmessage took {sw.Elapsed}");
-            messageMetadata.Logger?.LogInformation($"Deleted blob '{blobName}'");
+                sw.Restart();
+                messageMetadata.Logger?.LogInformation($"Deleting blob '{blobName}'");
+                await _blobClient.DeleteBlobAsync(blobName, cancellationToken);
+                messageMetadata.Logger?.LogInformation($"TIMETRACKING blobmessagehandler: deleting blobmessage took {sw.Elapsed}");
+                messageMetadata.Logger?.LogInformation($"Deleted blob '{blobName}'");
+            }
+            else
+            {
+                messageMetadata.Logger?.LogError($"Blob '{blobName}' not found");
+            }
         }
         else
         {
